feat: move config files to server/ and app/ directories
BREAKING: Config file locations changed:
- auth.config.ts → server/auth.config.ts
- client-auth.config.ts → app/auth.config.ts

New module options for custom paths:
- auth.serverConfig (default: 'server/auth.config')
- auth.clientConfig (default: 'app/auth.config')

Closes #2
