+++ docs/content/1.getting-started/0.index.md
@@ -81,7 +81,7 @@ NUXT_BETTER_AUTH_SECRET="generate-a-32-char-secret"
 
 ### Create Config Files
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 
 export default defineServerAuth(() => ({
@@ -89,7 +89,7 @@ export default defineServerAuth(() => ({
 }))
 ```
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { createAuthClient } from 'better-auth/vue'
 
 export function createAppAuthClient(baseURL: string) {
