+++ docs/content/1.getting-started/2.installation.md
@@ -78,12 +78,12 @@ export default defineNuxtConfig({
 
 Two configuration files connect Nuxt with Better Auth.
 
-1. **Server Configuration**: `auth.config.ts` (project root)
-2. **Client Configuration**: `client-auth.config.ts` (project root, or `app/` for Nuxt 4)
+1. **Server Configuration**: `server/auth.config.ts`
+2. **Client Configuration**: `app/auth.config.ts`
 
 Create these two configuration files:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 
 export default defineServerAuth(() => ({
@@ -91,7 +91,7 @@ export default defineServerAuth(() => ({
 }))
 ```
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { createAuthClient } from 'better-auth/vue'
 
 export function createAppAuthClient(baseURL: string) {
