+++ docs/content/1.getting-started/3.configuration.md
@@ -19,6 +19,18 @@ export default defineNuxtConfig({
 ```
 
 ::field-group
+  ::field{name="serverConfig" type="string"}
+    Default: `'server/auth.config'`
+
+    Path to the server auth config file, relative to the project root.
+  ::
+
+  ::field{name="clientConfig" type="string"}
+    Default: `'app/auth.config'`
+
+    Path to the client auth config file, relative to the project root.
+  ::
+
   ::field{name="redirects" type="{ login?: string, guest?: string }"}
     Default: `{ login: '/login', guest: '/' }`
 
@@ -38,13 +50,13 @@ export default defineNuxtConfig({
 
 ## Server Configuration
 
-Define your authentication logic in `auth.config.ts` (project root), including plugins, providers, and settings.
+Define your authentication logic in `server/auth.config.ts`, including plugins, providers, and settings.
 
 ### defineServerAuth
 
 Use the `defineServerAuth` helper to ensure type safety and access context.
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 
 export default defineServerAuth(() => {
@@ -66,7 +78,7 @@ The module automatically injects `secret` and `baseURL`. You don't need to confi
 
 The `defineServerAuth` callback receives a context object with useful properties:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 import { drizzleAdapter } from 'better-auth/adapters/drizzle'
