+++ docs/content/1.getting-started/4.client-setup.md
@@ -6,13 +6,13 @@ navigation.icon: i-lucide-monitor
 
 ## Create the Client Factory
 
-Create `client-auth.config.ts` in your project root (or `app/` for Nuxt 4) and export a `createAppAuthClient` function.
+Create `app/auth.config.ts` and export a `createAppAuthClient` function.
 
 ::note
 This factory function exists because the module needs to inject the correct `baseURL` at runtime. The module calls this function and provides the URL automatically.
 ::
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { createAuthClient } from 'better-auth/vue'
 
 export function createAppAuthClient(baseURL: string) {
@@ -22,9 +22,9 @@ export function createAppAuthClient(baseURL: string) {
 
 ## Using Plugins
 
-If you added a plugin in your server config (`auth.config.ts`), make sure to add its client equivalent here.
+If you added a plugin in your server config (`server/auth.config.ts`), make sure to add its client equivalent here.
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { createAuthClient } from 'better-auth/vue'
 import { adminClient } from 'better-auth/client/plugins'
 
@@ -42,7 +42,7 @@ export function createAppAuthClient(baseURL: string) {
 
 ### Admin + Two Factor
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { createAuthClient } from 'better-auth/vue'
 import { adminClient, twoFactorClient } from 'better-auth/client/plugins'
