+++ docs/content/1.getting-started/5.type-augmentation.md
@@ -9,7 +9,7 @@ When you add Better Auth plugins that extend the user or session objects, TypeSc
 ## Automatic Type Inference
 
 Types are automatically inferred when you:
-- Add plugins to `auth.config.ts`
+- Add plugins to `server/auth.config.ts`
 - Use the module's type exports
 
 No additional configuration is needed for most cases.
@@ -18,7 +18,7 @@ No additional configuration is needed for most cases.
 
 If you add the `admin` plugin:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { admin } from 'better-auth/plugins'
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 
@@ -65,7 +65,7 @@ The `#nuxt-better-auth` virtual module exports `AuthUser` and `AuthSession` inte
 **Types not updating?**
 1. Restart your IDE's TypeScript server
 2. Run `npx nuxi prepare` to regenerate types
-3. Check that your `auth.config.ts` has no syntax errors
+3. Check that your `server/auth.config.ts` has no syntax errors
 
 ::callout{icon="i-lucide-arrow-right" to="/getting-started/how-it-works"}
 Now that you are set up, learn **How it Works**.
