+++ docs/content/1.getting-started/6.schema-generation.md
@@ -13,7 +13,7 @@ Better Auth requires tables for users, sessions, accounts. This module creates t
 
 ## How it works
 
-The module analyzes your `auth.config.ts` at build time to determine which tables are required by your core configuration and enabled plugins (e.g., `twoFactor`, `passkey`).
+The module analyzes your `server/auth.config.ts` at build time to determine which tables are required by your core configuration and enabled plugins (e.g., `twoFactor`, `passkey`).
 
 1. **Analysis**: Reads `plugins` from your config.
 2. **Generation**: Generates a Drizzle schema file corresponding to your database dialect (SQLite, PostgreSQL, MySQL).
@@ -47,7 +47,7 @@ export default defineNuxtConfig({
 
 When you add a new plugin that requires database tables (e.g., adding `twoFactor`):
 
-1. Update `auth.config.ts`.
+1. Update `server/auth.config.ts`.
 2. Restart your development server.
 3. The module regenerates the schema.
 4. If using NuxtHub local development, the tables are ready.
