+++ docs/content/1.getting-started/7.how-it-works.md
@@ -8,10 +8,10 @@ navigation.icon: i-lucide-workflow
 
 ::steps
 ### 1. Initialization
-During build, the module reads `auth.config.ts` and generates the auth handler. If using NuxtHub, it also generates the database schema automatically.
+During build, the module reads `server/auth.config.ts` and generates the auth handler. If using NuxtHub, it also generates the database schema automatically.
 
 ### 2. Server Runtime
-The module injects a catch-all API handler at `/api/auth/**`. This handler processes all authentication requests (login, logout, session fetching) using the configuration provided in `auth.config.ts`.
+The module injects a catch-all API handler at `/api/auth/**`. This handler processes all authentication requests (login, logout, session fetching) using the configuration provided in `server/auth.config.ts`.
 
 ### 3. Client Runtime
 On the client side, `useUserSession` initializes the Better Auth client. It establishes a reactive connection to the session state, ensuring that your UI updates immediately when a user logs in or out.
