+++ docs/content/2.core-concepts/2.sessions.md
@@ -94,7 +94,7 @@ const { fetchSession, ready } = useUserSession()
 
 Configure session duration in your auth config:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 export default defineServerAuth(() => ({
   session: {
     cookie: {
