+++ docs/content/2.core-concepts/5.security-caveats.md
@@ -11,7 +11,7 @@ Client-side redirects improve UX but don't prevent unauthorized access. Always v
 
 Better Auth performs origin checks for cookie-based requests (it validates `Origin` / `Referer`). If you use multiple domains (custom domains, preview URLs, etc.), add them to `trustedOrigins` in your auth config.
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 export default defineServerAuth(() => ({
   trustedOrigins: [
     'http://localhost:3000',
