+++ docs/content/3.guides/1.role-based-access.md
@@ -5,7 +5,7 @@ description: Protect routes using generic field matching on AuthUser.
 
 ## With Admin Plugin
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { admin } from 'better-auth/plugins'
 export default defineServerAuth(() => ({ plugins: [admin()] }))
