+++ docs/content/3.guides/3.custom-database.md
@@ -15,7 +15,7 @@ Better Auth supports any database through adapters. Use this guide when you need
 
 ## Drizzle Adapter
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 import { drizzleAdapter } from 'better-auth/adapters/drizzle'
 import { drizzle } from 'drizzle-orm/node-postgres'
@@ -30,7 +30,7 @@ export default defineServerAuth(() => ({
 
 ## Prisma Adapter
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 import { prismaAdapter } from 'better-auth/adapters/prisma'
 import { PrismaClient } from '@prisma/client'
@@ -45,7 +45,7 @@ export default defineServerAuth(() => ({
 
 ## Kysely Adapter
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 import { kyselyAdapter } from 'better-auth/adapters/kysely'
 import { Kysely, PostgresDialect } from 'kysely'
