+++ docs/content/3.guides/3.oauth-providers.md
@@ -28,9 +28,9 @@ GOOGLE_CLIENT_ID="your-client-id"
 GOOGLE_CLIENT_SECRET="your-client-secret"
 ```
 
-### Configure `auth.config.ts`
+### Configure `server/auth.config.ts`
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 
 export default defineServerAuth(() => ({
@@ -73,7 +73,7 @@ Better Auth supports 20+ providers (Apple, Discord, Facebook, etc). The pattern 
 
 1. Create OAuth app, get client ID/secret
 2. Add `{PROVIDER}_CLIENT_ID` and `{PROVIDER}_CLIENT_SECRET` to `.env`
-3. Reference credentials in `auth.config.ts`
+3. Reference credentials in `server/auth.config.ts`
 4. Call `signIn.social({ provider: 'providerId' })`
 
 :read-more{to="https://www.better-auth.com/docs/authentication/other-social-providers" title="Full provider list"}
