+++ docs/content/3.guides/5.migrate-from-nuxt-auth-utils.md
@@ -56,7 +56,7 @@ export default defineOAuthGitHubEventHandler({
 
 ```ts
 // nuxt-better-auth - Declarative config
-// auth.config.ts
+// server/auth.config.ts
 export default defineServerAuth(() => ({
   socialProviders: {
     github: {
@@ -105,7 +105,7 @@ export default defineOAuthGitHubEventHandler({
 
 **After:**
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 export default defineServerAuth(() => ({
   socialProviders: {
     github: {
@@ -244,7 +244,7 @@ The simplest approach: require all users to reset their passwords during migrati
 
 Implement a custom password verifier that tries both formats:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { scrypt, timingSafeEqual } from 'node:crypto'
 
 async function verifyLegacyHash(password: string, hash: string): Promise<boolean> {
