+++ docs/content/3.guides/7.two-factor-auth.md
@@ -9,7 +9,7 @@ Better Auth ships a `twoFactor` plugin with TOTP, OTP, and backup codes. It requ
 
 ### 1) Enable the Plugin (Server)
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 import { defineServerAuth } from '@onmax/nuxt-better-auth'
 import { twoFactor } from 'better-auth/plugins/two-factor'
 
@@ -26,7 +26,7 @@ export default defineServerAuth(() => ({
 
 ### 2) Register Client Plugin
 
-```ts [client-auth.config.ts]
+```ts [app/auth.config.ts]
 import { navigateTo } from '#imports'
 import { createAuthClient } from 'better-auth/vue'
 import { twoFactorClient } from 'better-auth/client/plugins'
