+++ docs/content/4.integrations/1.nuxthub.md
@@ -140,7 +140,7 @@ See [NuxtHub Database docs](https://hub.nuxt.com/docs/features/database) for mor
 
 ## Schema Generation Details
 
-The module analyzes your `auth.config.ts` at build time:
+The module analyzes your `server/auth.config.ts` at build time:
 
 1. **Reads plugins** from your config
 2. **Generates Drizzle schema** matching your dialect
@@ -154,7 +154,7 @@ The module analyzes your `auth.config.ts` at build time:
 
 With NuxtHub, you get access to the Drizzle database instance in your server config:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 export default defineServerAuth(({ db }) => ({
   // db is the Drizzle instance from NuxtHub
 }))
