+++ docs/content/6.troubleshooting/1.faq.md
@@ -40,7 +40,7 @@ Yes. Better Auth supports any database with a Drizzle, Prisma, or Kysely adapter
 
 Configure in Better Auth, then augment types:
 
-```ts [auth.config.ts]
+```ts [server/auth.config.ts]
 export default defineServerAuth(() => ({
   user: {
     additionalFields: {
@@ -72,20 +72,20 @@ If using NuxtHub for database features, you need `@nuxthub/core` version **0.10.
 pnpm add @nuxthub/core@latest
 ```
 
-## "Missing auth.config.ts"
+## "Missing server/auth.config.ts"
 
 The module validates config files at build time.
-Create `auth.config.ts` in project root exporting `defineServerAuth(...)`.
+Create `server/auth.config.ts` exporting `defineServerAuth(...)`.
 
-## "Missing client-auth.config.ts"
+## "Missing app/auth.config.ts"
 
-The module expects a client config file at `client-auth.config.ts` (or `app/client-auth.config.ts` for Nuxt 4) exporting `createAppAuthClient(baseURL)`.
+The module expects a client config file at `app/auth.config.ts` exporting `createAppAuthClient(baseURL)`.
 
 ## "auth.config.ts validation failed"
 
 - Ensure the file is **default-exported** and wrapped with `defineServerAuth`.
 - Do not set `secret`/`baseURL` manually (the module injects these).
-- Confirm the file name is exactly `auth.config.ts` in your project root.
+- Confirm the file is at the expected path (default: `server/auth.config.ts`).
 
 ## No schema file is generated
