+++ test/fixtures/basic/app/auth.config.ts
