+++ test/fixtures/basic/server/auth.config.ts
@@ -1,4 +1,4 @@
-import { defineServerAuth } from '../../../src/runtime/config'
+import { defineServerAuth } from '../../../../src/runtime/config'
 
 export default defineServerAuth(() => ({
   appName: 'Test App',
