+++ test/fixtures/no-db/app/auth.config.ts
