+++ test/fixtures/no-db/server/auth.config.ts
@@ -1,7 +1,7 @@
-import { defineServerAuth } from '../../../src/runtime/config'
+import { defineServerAuth } from '../../../../src/runtime/config'
 
 export default defineServerAuth(() => ({
-  appName: 'No-Hub Test App',
+  appName: 'Database-less Test App',
   // OAuth only - email/password needs database
   socialProviders: {
     github: { clientId: 'test', clientSecret: 'test' },
