+++ test/fixtures/no-hub/app/auth.config.ts
