+++ test/fixtures/no-hub/server/auth.config.ts
@@ -1,7 +1,7 @@
-import { defineServerAuth } from '../../../src/runtime/config'
+import { defineServerAuth } from '../../../../src/runtime/config'
 
 export default defineServerAuth(() => ({
-  appName: 'Database-less Test App',
+  appName: 'No-Hub Test App',
   // OAuth only - email/password needs database
   socialProviders: {
     github: { clientId: 'test', clientSecret: 'test' },
